deployment/requirements.txt
torch
torchvision
streamlit
numpy
pillow
matplotlib
